requirements.txt
piexif==1.1.3
Pillow==9.5.0
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
shiboken6==6.5.1.1
